Datasource/Loki: Fixes regression where enhanceDataFrame was not called Closes #20642
